@@ -5,9 +5,8 @@
   "main": "build/index.js",
   "scripts": {
     "bump": "hi-bump",
-    "start": "scripts/watch.sh",
-    "test": "mocha --compilers js:babel/register --recursive test/*.js",
-    "test:watch": "mocha --compilers js:babel/register --recursive --watch test/*.js"
+    "start": "tsc",
+    "watch": "tsc -w"
   },
   "repository": {
     "type": "git",
@@ -20,35 +19,14 @@
   },
   "homepage": "https://github.com/Hire-Forms/hire-forms-options",
   "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-core": "^6.9.1",
-    "babel-eslint": "^6.0.4",
-    "babel-plugin-object-assign": "^1.2.1",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-1": "^6.5.0",
-    "babelify": "^7.3.0",
-    "derequire": "^2.0.3",
-    "eslint": "^2.11.1",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-import": "^1.8.1",
-    "eslint-plugin-jsx-a11y": "^1.3.0",
-    "eslint-plugin-react": "^5.1.1",
+    "@types/react": "^15.0.13",
     "hi-bump": "^1.1.1",
-    "jsdom": "^9.2.1",
-    "mocha": "^2.5.3",
-    "should": "^9.0.0",
-    "testdom": "^2.0.0",
-    "watchify": "^3.7.0"
+    "typescript": "^2.2.1"
   },
   "dependencies": {
-    "classlist-polyfill": "^1.0.3",
+    "classnames": "^2.2.5",
     "hire-forms-prop-types": "2.0.1",
-    "hire-forms-utils": "^1.2.0"
-  },
-  "peerDependencies": {
-    "classnames": "^2.1.2",
-    "react": "^15.0.2",
-    "react-dom": "^15.0.2"
+    "hire-forms-utils": "^1.2.0",
+    "react": "^15.4.2"
   }
 }
